+++ cedar-policy-core/src/entities/json/entities.rs
@@ -297,7 +297,7 @@ impl<'e, S: Schema> EntityJsonParser<'e, S> {
                         if parent_euid.is_action() {
                             Ok(())
                         } else {
-                            Err(JsonDeserializationError::ActionParentIsNonAction {
+                            Err(JsonDeserializationError::ActionParentIsNotAction {
                                 uid: uid.clone(),
                                 parent: parent_euid.clone(),
                             })